Revert "feat!: export crypto primitives (#1728)"
This reverts commit 7eb3375f
