+++ packages/message-encryption/src/waku_payload.spec.ts
@@ -9,9 +9,9 @@ import {
   encryptSymmetric,
   postCipher,
   preCipher
-} from "./encryption.js";
+} from "./waku_payload.js";
 
-describe("Waku Encryption", function () {
+describe("Waku Payload", function () {
   this.timeout(20000);
   it("Asymmetric encrypt & decrypt", async function () {
     await fc.assert(
