+++ packages/tests/tests/ephemeral.node.spec.ts
@@ -7,15 +7,13 @@ import {
 import { IFilterSubscription, Protocols } from "@waku/interfaces";
 import type { LightNode } from "@waku/interfaces";
 import {
-  generatePrivateKey,
-  generateSymmetricKey,
-  getPublicKey
-} from "@waku/message-encryption";
-import {
   createDecoder as eciesDecoder,
-  createEncoder as eciesEncoder
+  createEncoder as eciesEncoder,
+  generatePrivateKey,
+  getPublicKey
 } from "@waku/message-encryption/ecies";
 import {
+  generateSymmetricKey,
   createDecoder as symDecoder,
   createEncoder as symEncoder
 } from "@waku/message-encryption/symmetric";
