+++ packages/tests/tests/store/index.node.spec.ts
@@ -7,17 +7,15 @@ import {
 import type { IMessage, LightNode } from "@waku/interfaces";
 import { Protocols } from "@waku/interfaces";
 import {
-  generatePrivateKey,
-  generateSymmetricKey,
-  getPublicKey
-} from "@waku/message-encryption";
-import {
   createDecoder as createEciesDecoder,
-  createEncoder as createEciesEncoder
+  createEncoder as createEciesEncoder,
+  generatePrivateKey,
+  getPublicKey
 } from "@waku/message-encryption/ecies";
 import {
   createDecoder as createSymDecoder,
-  createEncoder as createSymEncoder
+  createEncoder as createSymEncoder,
+  generateSymmetricKey
 } from "@waku/message-encryption/symmetric";
 import { bytesToUtf8, utf8ToBytes } from "@waku/utils/bytes";
 import { expect } from "chai";
